test(conf): move stubs repo fixtures to conf.
