+++ tests/conftest.py
@@ -1,7 +1,6 @@
 """ Common Pytest Fixtures"""
 
 import importlib
-import json
 import shutil
 from pathlib import Path
 from pprint import PrettyPrinter
@@ -28,7 +27,6 @@ def cleanup_data(mocker):
     mocker.resetall()
     try:
         micropy.utils.ensure_valid_url.clear_cache()
-        micropy.stubs.source.StubRepo.repos = set()
     except Exception:
         importlib.reload(micropy)
 
@@ -171,17 +169,85 @@ def test_archive(shared_datadir):
     file_obj.close()
 
 
+micropy_source = micropy.stubs.repository_info.RepositoryInfo(
+    name="BradenM/micropy-stubs",
+    display_name="micropy-stubs",
+    source="https://my-mocked-source.com/bradenm",
+)
+
+micropython_source = micropy.stubs.repository_info.RepositoryInfo(
+    name="Josverl/micropython-stubs",
+    display_name="micropython-stubs",
+    # source="https://raw.githubusercontent.com/Josverl/micropython-stubs/main/publish/package_data.jsondb",
+    source="https://my-mocked-source.com/josverl",
+)
+
+
 @pytest.fixture
-def test_repo(test_urls, shared_datadir, mocker):
-    micropy.stubs.source.StubRepo.repos = set()
-    mocker.patch.object(micropy.stubs.source.utils, "is_downloadable", return_value=True)
-    mocker.patch.object(
-        micropy.stubs.source.utils, "ensure_valid_url", return_value=test_urls["valid"]
+def mock_manifests(mocker, requests_mock):
+    micropy_manifest = {
+        "name": "Micropy Stubs",
+        "location": "https://codeload.github.com/BradenM/micropy-stubs",
+        "source": "https://raw.githubusercontent.com/bradenm/micropy-stubs/source.json",
+        "path": "legacy.tar.gz/pkg/",
+        "packages": [
+            {
+                "name": "micropython",
+                "type": "firmware",
+                "sha256sum": "7ff2cce0237268cd52164b77b6c2df6be6249a67ee285edc122960af869b8ed2",
+            },
+            {"name": "esp8266-micropython-1.15.0", "type": "device", "sha256sum": "abc123"},
+        ],
+    }
+    micropython_manifest = {
+        "version": 2,
+        "keys": [
+            "description",
+            "hash",
+            "mpy_version",
+            "name",
+            "path",
+            "pkg_version",
+            "publish",
+            "stub_hash",
+            "stub_sources",
+        ],
+        "data": {
+            "160521968180811532": {
+                "name": "micropython-esp32-stubs",
+                "mpy_version": "1.18",
+                "publish": True,
+                "pkg_version": "1.18.post1",
+                "path": "publish/micropython-v1_18-esp32-stubs",
+                "stub_sources": [
+                    ["Firmware stubs", "stubs/micropython-v1_18-esp32"],
+                    ["Frozen stubs", "stubs/micropython-v1_18-frozen/esp32/GENERIC"],
+                    ["Core Stubs", "stubs/cpython_core-pycopy"],
+                ],
+                "description": "MicroPython stubs",
+                "hash": "712ebd85140b078ce6d9d3cbb9d7ffc18cf10aef",
+                "stub_hash": "",
+            }
+        },
+    }
+
+    requests_mock.get(
+        micropy_source.source,
+        json=micropy_manifest,
     )
-    test_data = json.loads((shared_datadir / "test_repo.json").read_text())
-    mock_get = mocker.patch.object(micropy.stubs.source.requests, "get")
-    mock_get.return_value.json.return_value = test_repo
-    repo = micropy.stubs.source.StubRepo(**test_data)
+    requests_mock.get(
+        micropython_source.source,
+        json=micropython_manifest,
+    )
+
+
+@pytest.fixture
+def test_repo(mock_manifests):
+    micropy_source.fetch_source.clear_cache()
+    micropython_source.fetch_source.clear_cache()
+    repo = micropy.stubs.repo.StubRepository()
+    repo = repo.add_repository(micropy_source)
+    repo = repo.add_repository(micropython_source)
     return repo
